Fixed a bug in BFS
The result of aab for the RegEx ab was "b" before, since the algorithm always deleted the char one was currently at if he had no way to continue from the current point. Now it checks whether it could start from this point and then differs between these two cases.